Account for presence of byte-order marker (BOM)
Tricky bug: The Unicode byte-order marker is present when Excel
saves a spreadsheet as CSV. However, data.csv erroneously loads
the BOM as part of the cell. This leads to bizarre behavior like
(not (= "ID" "ID")), only visible if you get the hex bytes of the
rendered string. Apache Commons IO deals with this.